loading widget and empty animation added
